requirements.txt
# Core
streamlit>=1.36.0
langchain>=0.2.10
langchain-community>=0.2.10
langchain-openai>=0.1.17

# Vector stores
faiss-cpu>=1.8.0
chromadb>=0.5.0

# Embeddings + tokenization
tiktoken>=0.7.0

# HTTP and parsing
requests>=2.32.3
httpx>=0.27.0
beautifulsoup4>=4.12.3

# Data & viz
pandas>=2.2.2
numpy>=1.26.4
plotly>=5.22.0

# Config
python-dotenv>=1.0.1
pydantic>=2.7.3

# Retry
tenacity>=8.3.0

# Tests
pytest>=8.3.2

# Optional OSS LLM via Gradio
gradio_client>=1.3.0

# JS eval for client IP capture in Streamlit
streamlit_js_eval>=0.1.7